+++ src/test/java/workouts/WorkoutListTest.java
@@ -189,8 +189,8 @@ class WorkoutListTest {
     @Test
     void deleteGym_properList_listOfSizeOne() throws CustomExceptions.InvalidInput, CustomExceptions.OutOfBounds {
         Gym gym1 = new Gym();
-        ArrayList<Integer> array1 = new ArrayList<>(Arrays.asList(1));
-        ArrayList<Integer> array2 = new ArrayList<>(Arrays.asList(1,2));
+        ArrayList<Double> array1 = new ArrayList<>(Arrays.asList(1.0));
+        ArrayList<Double> array2 = new ArrayList<>(Arrays.asList(1.0,2.0));
 
         gym1.addStation("Bench Press", 1, 50, array1);
         gym1.addStation("Shoulder Press", 2, 10, array2);
@@ -224,8 +224,8 @@ class WorkoutListTest {
     @Test
     void deleteGym_properListInvalidIndex_throwOutOfBoundsForRun() throws CustomExceptions.InvalidInput {
         Gym gym1 = new Gym();
-        ArrayList<Integer> array1 = new ArrayList<>(Arrays.asList(1));
-        ArrayList<Integer> array2 = new ArrayList<>(Arrays.asList(1,2));
+        ArrayList<Double> array1 = new ArrayList<>(Arrays.asList(1.0));
+        ArrayList<Double> array2 = new ArrayList<>(Arrays.asList(1.0,2.0));
         gym1.addStation("Bench Press", 1, 50, array1);
         gym1.addStation("Shoulder Press", 2, 10, array2);
         int invalidIndex = 5;
